lab06/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
